Extract to component
