+++ src/pages/ConditionDetails/Contents.tsx
@@ -1,10 +1,9 @@
-import React, { useCallback, useMemo, useState } from 'react'
+import React, { useCallback, useMemo } from 'react'
 import { NavLink } from 'react-router-dom'
 import styled from 'styled-components'
 
 import { ButtonCopy } from 'components/buttons/ButtonCopy'
 import { ButtonDropdownCircle } from 'components/buttons/ButtonDropdownCircle'
-import { ButtonExpand } from 'components/buttons/ButtonExpand'
 import { CenteredCard } from 'components/common/CenteredCard'
 import {
   Dropdown,
@@ -13,7 +12,7 @@ import {
   DropdownPosition,
 } from 'components/common/Dropdown'
 import { DisplayTablePositions } from 'components/form/DisplayTablePositions'
-import { DisplayHashesTableModal } from 'components/modals/DisplayHashesTableModal'
+import { OmenMarketsItem } from 'components/form/OmenMarketsItem'
 import { ExternalLink } from 'components/navigation/ExternalLink'
 import { FlexRow } from 'components/pureStyledComponents/FlexRow'
 import { Pill, PillTypes } from 'components/pureStyledComponents/Pill'
@@ -21,18 +20,16 @@ import { Row } from 'components/pureStyledComponents/Row'
 import { StripedList, StripedListItem } from 'components/pureStyledComponents/StripedList'
 import { FormatHash } from 'components/text/FormatHash'
 import { TitleValue } from 'components/text/TitleValue'
-import { INFORMATION_NOT_AVAILABLE, OMEN_URL_DAPP } from 'config/constants'
+import { INFORMATION_NOT_AVAILABLE } from 'config/constants'
 import { Web3ContextStatus, useWeb3ConnectedOrInfura } from 'contexts/Web3Context'
 import { useIsConditionFromOmen } from 'hooks/useIsConditionFromOmen'
 import { useLocalStorage } from 'hooks/useLocalStorageValue'
-import { useOmenMarkets } from 'hooks/useOmenMarkets'
 import { usePositions } from 'hooks/usePositions'
 import { useQuestion } from 'hooks/useQuestion'
 import { GetCondition_condition } from 'types/generatedGQLForCTE'
 import {
   formatTS,
   getConditionTypeTitle,
-  getOmenMarketURL,
   getRealityQuestionUrl,
   truncateStringInTheMiddle,
 } from 'util/tools'
@@ -67,28 +64,7 @@ export const Contents: React.FC<Props> = ({ condition }) => {
     resolved,
   } = condition
 
-  const {
-    data: dataOmenMarkets,
-    error: errorOmenMarkets,
-    loading: loadingOmenMarkets,
-  } = useOmenMarkets([conditionId])
-  const [openOmenMarkets, setOpenOmenMarkets] = useState(false)
-
   const isConnected = useMemo(() => status === Web3ContextStatus.Connected, [status])
-  const omenMarkets = useMemo(() => {
-    if (
-      !loadingOmenMarkets &&
-      !errorOmenMarkets &&
-      dataOmenMarkets?.condition?.fixedProductMarketMakers &&
-      dataOmenMarkets?.condition?.fixedProductMarketMakers?.length > 0
-    ) {
-      return dataOmenMarkets.condition.fixedProductMarketMakers
-    } else {
-      return []
-    }
-  }, [dataOmenMarkets, errorOmenMarkets, loadingOmenMarkets])
-
-  const areOmenMarketsMoreThanOne = useMemo(() => omenMarkets.length > 1, [omenMarkets])
 
   const dropdownItems = useMemo(() => {
     return [
@@ -250,37 +226,7 @@ export const Contents: React.FC<Props> = ({ condition }) => {
           }
         />
       </Row>
-      {omenMarkets.length > 0 && (
-        <Row>
-          <TitleValue
-            title={areOmenMarketsMoreThanOne ? 'Omen Markets' : 'Omen Market'}
-            value={
-              <FlexRow>
-                {omenMarkets[0].question ? omenMarkets[0].question.title : omenMarkets[0].id}
-                {!areOmenMarketsMoreThanOne && <ButtonCopy value={omenMarkets[0].id} /> && (
-                  <ExternalLink href={getOmenMarketURL(omenMarkets[0].id)} />
-                )}
-                {areOmenMarketsMoreThanOne && (
-                  <ButtonExpand onClick={() => setOpenOmenMarkets(true)} />
-                )}
-              </FlexRow>
-            }
-          />
-
-          {openOmenMarkets && omenMarkets.length > 0 && (
-            <DisplayHashesTableModal
-              hashes={omenMarkets.map(({ id, question }) => {
-                return { hash: id, title: question?.title || undefined }
-              })}
-              isOpen={openOmenMarkets}
-              onRequestClose={() => setOpenOmenMarkets(false)}
-              title="Omen Markets"
-              titleTable="Market Name"
-              url={OMEN_URL_DAPP}
-            />
-          )}
-        </Row>
-      )}
+      <OmenMarketsItem conditionsIds={[conditionId]} />
       <Row cols="1fr">
         <TitleValue
           title={"Condition's split positions"}
